requirements.txt
fastapi~=0.104.0
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
uvicorn>=0.24.0,<0.25.0